Created a separate controller and moved the palindrome URL handler inside it.
