+++ src/main/java/com/dinesh/palindromeapp/PalindromicApplication.java
@@ -1,0 +1,13 @@
+package com.dinesh.palindromeapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PalindromicApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PalindromicApplication.class, args);
+	}
+
+}
