--- src/test/java/com/dinesh/palindromeapp/PalindromeappApplicationTests.java
@@ -1,16 +1,0 @@
-package com.dinesh.palindromeapp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-@SpringBootTest
-class PalindromeappApplicationTests {
-
-	@Test
-	void IsUnitTestWorking() {
-		assertTrue(true);
-	}
-
-}
